feat: Update Histogram interface
Handle exceptions that occasionally are thrown by
com.dynatrace.dynahist.Histogram

+++ src/test/java/org/obd/metrics/api/DataGeneratorTest.java
@@ -54,7 +54,7 @@ public class DataGeneratorTest {
 
 		Histogram histogram = workflow.getDiagnostics().histogram().findBy(pid8l);
 		Assertions.assertThat(histogram.getMax()).isGreaterThan(histogram.getMin());
-		Assertions.assertThat(histogram.getMin()).isLessThan((long) histogram.getMean());
+		Assertions.assertThat(histogram.getMin()).isLessThan(histogram.getMean());
 		Assertions.assertThat(histogram.getMean()).isLessThan(histogram.getMax()).isGreaterThan(histogram.getMin());
 	}
 
@@ -97,7 +97,7 @@ public class DataGeneratorTest {
 		Histogram histogram = workflow.getDiagnostics().histogram().findBy(pid8l);
 
 		Assertions.assertThat(histogram.getMax()).isGreaterThan(histogram.getMin());
-		Assertions.assertThat(histogram.getMin()).isLessThan((long) histogram.getMean());
+		Assertions.assertThat(histogram.getMin()).isLessThan(histogram.getMean());
 		Assertions.assertThat(histogram.getMean()).isLessThan(histogram.getMax()).isGreaterThan(histogram.getMin());
 	}
